cleanup/refactor stats in detail. reduce clutter
