+++ package.json
@@ -19,6 +19,7 @@
         "axios": "^1.7.2",
         "dayjs": "^1.11.11",
         "echarts": "^5.5.0",
+        "file-saver": "^2.0.5",
         "localforage": "^1.10.0",
         "match-sorter": "^6.3.4",
         "papaparse": "^5.4.1",
