+++ src/pages/statistics/index.tsx
@@ -52,48 +52,37 @@ const Statistics: React.FC = () => {
                     handleYearSelect={handleYearSelect}
                 />
 
-                <Box
-                    sx={{
-                        flexGrow: 1,
-                        overflow: 'auto',
-                        display: 'flex',
-                        flexDirection: 'column',
-                        alignItems: 'center',
-                        padding: '2rem',
-                    }}
-                >
-                    {selectedTab === 'monthly' && <StatisticsTable year={selectedYear} />}
-                    {selectedTab === 'overall' && (
-                        <>
-                            {selectedChart && (
-                                <Paper
-                                    elevation={16}
-                                    sx={{
-                                        height: '100%',
-                                        width: '100%',
-                                        display: 'flex',
-                                        flexDirection: 'column',
-                                        alignItems: 'center',
-                                        justifyContent: 'center',
-                                        // backgroundColor: '#f5f5f5',
-                                        backgroundColor: 'white',
-                                        borderRadius: '1.5rem',
-                                    }}
-                                >
-                                    {selectedChart === 'JVMs' ? (
-                                        <JVMChart title={chartTitles[selectedChart]} />
-                                    ) : (
-                                        <Chart
-                                            key={`${selectedChart}-${sidebarOpen}`} // Force re-render on sidebar open/close
-                                            csvPath={`https://raw.githubusercontent.com/jenkins-infra/infra-statistics/gh-pages/jenkins-stats/svg/total-${selectedChart}.csv`}
-                                            title={chartTitles[selectedChart]}
-                                        />
-                                    )}
-                                </Paper>
-                            )}
-                        </>
-                    )}
-                </Box>
+                {selectedTab === 'monthly' && <StatisticsTable year={selectedYear} />}
+                {selectedTab === 'overall' && (
+                    <>
+                        {selectedChart && (
+                            <Paper
+                                elevation={16}
+                                sx={{
+                                    width: '100%',
+                                    display: 'flex',
+                                    flexDirection: 'column',
+                                    justifyContent: 'center',
+                                    alignItems: 'center',
+                                    padding: '3rem',
+                                    margin: '2rem',
+                                    backgroundColor: 'white',
+                                    borderRadius: 5,
+                                }}
+                            >
+                                {selectedChart === 'JVMs' ? (
+                                    <JVMChart title={chartTitles[selectedChart]} />
+                                ) : (
+                                    <Chart
+                                        key={`${selectedChart}-${sidebarOpen}`} // Force re-render on sidebar open/close
+                                        csvPath={`https://raw.githubusercontent.com/jenkins-infra/infra-statistics/gh-pages/jenkins-stats/svg/total-${selectedChart}.csv`}
+                                        title={chartTitles[selectedChart]}
+                                    />
+                                )}
+                            </Paper>
+                        )}
+                    </>
+                )}
             </Box>
         </Box>
     )
